spine/requirements.txt
python-docx>=1.1.0
pytest>=8.0.0
langchain>=0.1.0
chromadb>=0.4.0
lxml>=5.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
